upgrade oc-common and handle make_model in a better way.
