+++ onnxconverter_common/topology.py
@@ -3,26 +3,16 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 ###############################################################################
-
 import re
 import warnings
-from logging import getLogger
 from distutils.version import StrictVersion
 from onnx import helper
-from .metadata_props import add_metadata_props
 from . import registration
-from . import utils
 from .data_types import *
+from .onnx_ex import make_model_ex, OPSET_TO_IR_VERSION, DEFAULT_OPSET_NUMBER
 from .container import ModelComponentContainer
 from .optimizer import optimize_onnx
 from .interface import OperatorBase, ScopeBase
-
-OPSET_TO_IR_VERSION = {
-    1: 3, 2: 3, 3: 3, 4: 3, 5: 3, 6: 3,
-    7: 3, 8: 4, 9: 4, 10: 5, 11: 6, 12: 7
-}
-
-DEFAULT_OPSET_NUMBER = 11  # The maximum opset supported by the converter in the code branch.
 
 
 class Variable:
@@ -817,9 +807,6 @@ def convert_topology(topology, model_name, doc_string, target_opset, targeted_on
     # Add extra information related to the graph
     graph.value_info.extend(container.value_info)
 
-    # Create model
-    onnx_model = helper.make_model(graph)
-
     # Merge operator sets for the same domain, the largest version number would be kept
     purified_operator_set = dict()
     for op_domain, op_version in container.node_domain_version_pair_sets:
@@ -828,43 +815,6 @@ def convert_topology(topology, model_name, doc_string, target_opset, targeted_on
         else:
             purified_operator_set[op_domain] = max(purified_operator_set[op_domain], op_version)
 
-    # Fill operator sets
-    i = 0
-    for op_domain, op_version in purified_operator_set.items():
-        if i == 0 and len(onnx_model.opset_import) == 1:
-            # Overwrite the default operator set created by helper.make_model(...)
-            op_set = onnx_model.opset_import[0]
-        else:
-            # Just create one ONNX element in opset_import
-            op_set = onnx_model.opset_import.add()
-        op_set.domain = op_domain
-        op_set.version = op_version
-        i += 1
-        if container.target_opset < op_version:
-            raise RuntimeError(('The specified opset %d is too low to convert this model, ' +
-                                'which requires at least opset %d.') % (container.target_opset, op_version))
-        elif container.target_opset > op_version:
-            getLogger('onnxmltools').warning('The maximum opset needed by this model is only %d.' % op_version)
-
-    # Add extra information
-    add_metadata_props(onnx_model, topology.metadata_props, container.target_opset)
-    opv = _get_main_opset_version(onnx_model) or container.target_opset
-    irv = OPSET_TO_IR_VERSION.get(opv, onnx_proto.IR_VERSION)
-    onnx_model.ir_version = irv
-    onnx_model.producer_name = utils.get_producer()
-    onnx_model.producer_version = utils.get_producer_version()
-    onnx_model.domain = utils.get_domain()
-    onnx_model.model_version = utils.get_model_version()
-    onnx_model.doc_string = doc_string
-
+    onnx_model = make_model_ex(graph, purified_operator_set,
+                               target_opset, topology.metadata_props, doc_string=doc_string)
     return onnx_model
-
-
-def _get_main_opset_version(model):
-    """
-    Returns the main opset version.
-    """
-    for op in model.opset_import:
-        if op.domain == '':
-            return op.version
-    return None
